feat: success getting working apic

+++ kernel/dev/cpu/apic/apic.h
@@ -38,7 +38,7 @@
 int apic_is_supported();
 void apic_setup(void);
 void apic_eoi(void);
-void apic_write(uint8_t reg, uint32_t value);
-uint32_t apic_read(uint8_t reg);
+void apic_write(uint32_t reg, uint32_t value);
+uint32_t apic_read(uint32_t reg);
 
 #endif
